#9 in progress. Separated ServerMaster and ServerConfig. Splitting server blocks into elements in progress.
