+++ Makefile
@@ -6,15 +6,16 @@
 #    By: aulicna <aulicna@student.42.fr>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/01/13 11:40:54 by aulicna           #+#    #+#              #
-#    Updated: 2024/05/22 18:37:16 by aulicna          ###   ########.fr        #
+#    Updated: 2024/05/29 12:29:26 by aulicna          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 NAME = webserv
 
 SRC = main.cpp \
-	Server.cpp \
+	ServerMaster.cpp \
 	ServerConfig.cpp \
+	Server.cpp \
 	Client.cpp
 
 OBJ = $(SRC:.cpp=.o)
