+++ ServerConfig.hpp
@@ -5,35 +5,26 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: aulicna <aulicna@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/05/22 18:28:00 by aulicna           #+#    #+#             */
-/*   Updated: 2024/05/28 15:32:52 by aulicna          ###   ########.fr       */
+/*   Created: 2024/05/29 12:21:17 by aulicna           #+#    #+#             */
+/*   Updated: 2024/05/29 14:50:09 by aulicna          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef SERVERCONFIG_HPP
 # define SERVERCONFIG_HPP
 
-# include "webserv.hpp"
+# include <iostream>
 
 class ServerConfig
 {
 	public:
-		ServerConfig(std::string configFile);
+		ServerConfig(std::string &serverBlock);
 		ServerConfig(const ServerConfig& copy);
 		ServerConfig	&operator=(const ServerConfig &src);
 		~ServerConfig(void);
 
-		std::string	getFileContent(void) const;
-
-		void	removeCommentsAndEmptyLines(void);
-		void	detectServerBlocks(void);
-
-		void	printServerBlocks(void) const;
-	
 	private:
 		ServerConfig(void);
-		std::string					_configContent;
-		std::vector<std::string>	_serverBlocks;
 
 };
