+++ config_files/default.conf
@@ -1,6 +1,5 @@
 
 
-
 server {
 
     listen 8002;
